+++ runtime/android/java/src/org/xwalk/runtime/XWalkRuntimeViewProvider.java
@@ -15,58 +15,31 @@ import org.xwalk.runtime.extension.XWalkExtensionContextImpl;
 import org.xwalk.runtime.extension.XWalkExtensionManager;
 
 /**
- * The abstract class to provide the bridge between XWalkRuntimeView and the
+ * The interface to provide the bridge between XWalkRuntimeView and the
  * real implementation like runtime core.
  *
  * This class is also used by XWalkExtensionManager to get the capability to
  * build runtime extension system.
  */
-public abstract class XWalkRuntimeViewProvider {
-    private Context mContext;
-    private Activity mActivity;
-    private XWalkExtensionManager mExtensionManager;
-    private XWalkExtensionContextImpl mExtensionContext;
+public interface XWalkRuntimeViewProvider {
+    // For the embedded mode, the two arguments are the same.
+    public void init(Context context, Activity activity);
+    public XWalkExtensionContext getExtensionContext();
 
-    XWalkRuntimeViewProvider(Context context, Activity activity) {
-        mContext = context;
-        mActivity = activity;
-        mExtensionManager = new XWalkExtensionManager(context, activity, this);
-    }
+    // For handling life cycle and activity result.
+    public void onCreate();
+    public void onResume();
+    public void onPause();
+    public void onDestroy();
+    public void onActivityResult(int requestCode, int resultCode, Intent data);
 
-    public void init(Context context, Activity activity) {
-        mExtensionManager.loadExtensions();
-    }
-
-    public XWalkExtensionContext getExtensionContext() {
-        return mExtensionManager.getExtensionContext();
-    }
-
-    public void onCreate() {
-
-    }
-
-    public void onResume() {
-        mExtensionManager.onResume();
-    }
-
-    public void onPause() {
-        mExtensionManager.onPause();
-    }
-
-    public void onDestroy() {
-        mExtensionManager.onDestroy();
-    }
-
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        mExtensionManager.onActivityResult(requestCode, resultCode, data);
-    }
-
-    public abstract String getVersion();
-    public abstract View getView();
-    public abstract void loadAppFromUrl(String url);
-    public abstract void loadAppFromManifest(String manifestUrl);
-    public abstract String enableRemoteDebugging(String frontEndUrl, String socketName);
-    public abstract void disableRemoteDebugging();
+    // For RuntimeView APIs.
+    public String getVersion();
+    public View getView();
+    public void loadAppFromUrl(String url);
+    public void loadAppFromManifest(String manifestUrl);
+    public String enableRemoteDebugging(String frontEndUrl, String socketName);
+    public void disableRemoteDebugging();
 
     /*
      * Once runtime extension is created, notify the internal mechanism so that
@@ -79,16 +52,12 @@ public abstract class XWalkRuntimeViewProvider {
      * Pass messages from native extension system to runtime extension system.
      * Might be overrided to do customizations here.
      */
-    public void onMessage(XWalkExtension extension, int instanceID, String message) {
-        extension.onMessage(instanceID, message);
-    }
+    public void onMessage(XWalkExtension extension, int instanceID, String message);
 
     /**
      * Pass synchronized messages.
      */
-    public String onSyncMessage(XWalkExtension extension, int instanceID, String message) {
-        return extension.onSyncMessage(instanceID, message);
-    }
+    public String onSyncMessage(XWalkExtension extension, int instanceID, String message);
 
     /**
      * Pass message from runtime extension system to native and then to JavaScript.
@@ -102,6 +71,6 @@ public abstract class XWalkRuntimeViewProvider {
     public abstract void broadcastMessage(XWalkExtension extension, String message);
 
     // For instrumentation test.
-    public abstract String getTitleForTest();
-    public abstract void setCallbackForTest(Object callback);
+    public String getTitleForTest();
+    public void setCallbackForTest(Object callback);
 }
